# Please enter the commit message for your changes. Lines starting # with '#' will be ignored, and an empty message aborts the commit. # # On branch qinqiangxu/remove_dup_ts_test_cases # Cherry-pick currently in progress. # # Changes to be committed: #	modified:   packages/tests/scripts/pvt.json #	modified:   packages/tests/scripts/sampleCaseCreate.ts #	renamed:    packages/tests/src/ui-test/remotedebug/remotedebug-chatdata-customapi-ts-azureopenai-win-only.test.ts -> packages/tests/src/unused-ui-test/remotedebug/remotedebug-chatdata-customapi-ts-azureopenai-win-only.test.ts #	renamed:    packages/tests/src/ui-test/remotedebug/remotedebug-chatdata-customize-js-azureopenai-win-only.test.ts -> packages/tests/src/unused-ui-test/remotedebug/remotedebug-chatdata-customize-js-azureopenai-win-only.test.ts #	renamed:    packages/tests/src/ui-test/remotedebug/remotedebug-tab-nosso-ts-win-only.test.ts -> packages/tests/src/unused-ui-test/remotedebug/remotedebug-tab-nosso-ts-win-only.test.ts #	renamed:    packages/tests/src/ui-test/samples/sample-bot-sso.test.ts -> packages/tests/src/unused-ui-test/samples/sample-bot-sso.test.ts #	renamed:    packages/tests/src/ui-test/samples/sample-hello-world-tab-with-backend.test.ts -> packages/tests/src/unused-ui-test/samples/sample-hello-world-tab-with-backend.test.ts #	renamed:    packages/tests/src/ui-test/samples/sample-intelligent-data-chart.test.ts -> packages/tests/src/unused-ui-test/samples/sample-intelligent-data-chart.test.ts #	renamed:    packages/tests/src/ui-test/samples/sample-large-scale-notification.test.ts -> packages/tests/src/unused-ui-test/samples/sample-large-scale-notification.test.ts #	renamed:    packages/tests/src/ui-test/samples/sample-reddit-link.test.ts -> packages/tests/src/unused-ui-test/samples/sample-reddit-link.test.ts #	renamed:    packages/tests/src/ui-test/templates/basic-bot-ts-localdebug.test.ts -> packages/tests/src/unused-ui-test/templates/basic-bot-ts-localdebug.test.ts #	renamed:    packages/tests/src/ui-test/templates/basic-bot-ts-remotedebug.test.ts -> packages/tests/src/unused-ui-test/templates/basic-bot-ts-remotedebug.test.ts #	renamed:    packages/tests/src/ui-test/treeview/treeview-invalidname.test.ts -> packages/tests/src/unused-ui-test/treeview/treeview-invalidname.test.ts #	renamed:    packages/tests/src/ui-test/treeview/treeview-newproject-in-exist-project.test.ts -> packages/tests/src/unused-ui-test/treeview/treeview-newproject-in-exist-project.test.ts #
