+++ packages/tests/src/unused-ui-test/remotedebug/remotedebug-chatdata-customapi-ts-azureopenai-win-only.test.ts
@@ -2,24 +2,21 @@
 // Licensed under the MIT license.
 
 /**
- * @author Helly Zhang <v-helzha@microsoft.com>
+ * @author Anne Fu <v-annefu@microsoft.com>
  */
 import * as path from "path";
 import { VSBrowser } from "vscode-extension-tester";
-import { Timeout, ValidationContent } from "../../utils/constants";
+import { Lang, Timeout, ValidationContent } from "../../utils/constants";
 import {
   RemoteDebugTestContext,
   provisionProject,
   deployProject,
-} from "./remotedebugContext";
+} from "../../ui-test/remotedebug/remotedebugContext";
 import {
   execCommandIfExist,
   createNewProject,
 } from "../../utils/vscodeOperation";
-import {
-  initPage,
-  validateWelcomeAndReplyBot,
-} from "../../utils/playwrightOperation";
+import { initPage, validateCustomapi } from "../../utils/playwrightOperation";
 import { Env, OpenAiKey } from "../../utils/env";
 import { it } from "../../utils/it";
 import { editDotEnvFile, validateFileExist } from "../../utils/commonUtils";
@@ -36,7 +33,7 @@ describe("Remote debug Tests", function () {
   beforeEach(async function () {
     // ensure workbench is ready
     this.timeout(Timeout.prepareTestCase);
-    remoteDebugTestContext = new RemoteDebugTestContext("chatdata");
+    remoteDebugTestContext = new RemoteDebugTestContext("cdcustomapi");
     testRootFolder = remoteDebugTestContext.testRootFolder;
     appName = remoteDebugTestContext.appName;
     newAppFolderName = appName + appNameCopySuffix;
@@ -61,18 +58,19 @@ describe("Remote debug Tests", function () {
   });
 
   it(
-    "[auto][JS][Azure OpenAI] Remote debug for basic rag bot using customize data",
+    "[auto][TS][Azure OpenAI] Remote debug for Custom Copilot Rag Custom Api",
     {
-      testPlanCaseId: 27569147,
-      author: "v-helzha@microsoft.com",
+      testPlanCaseId: 28891618,
+      author: "v-annefu@microsoft.com",
     },
     async function () {
       const driver = VSBrowser.instance.driver;
-      await createNewProject("chatdata", appName, {
+      await createNewProject("cdcustomapi", appName, {
         aiType: "Azure OpenAI",
-        dataOption: "Customize",
+        lang: Lang.TS,
+        dataOption: "Custom API",
       });
-      validateFileExist(projectPath, "src/index.js");
+      validateFileExist(projectPath, "src/index.ts");
       const envPath = path.resolve(projectPath, "env", ".env.dev.user");
       const isRealKey = OpenAiKey.azureOpenAiKey ? true : false;
       const azureOpenAiKey = OpenAiKey.azureOpenAiKey
@@ -89,7 +87,7 @@ describe("Remote debug Tests", function () {
       editDotEnvFile(envPath, "AZURE_OPENAI_ENDPOINT", azureOpenAiEndpoint);
       editDotEnvFile(
         envPath,
-        "AZURE_OPENAI_DEPLOYMENT_NAME",
+        "AZURE_OPENAI_MODEL_DEPLOYMENT_NAME",
         azureOpenAiModelDeploymentName
       );
       await provisionProject(appName, projectPath);
@@ -111,16 +109,16 @@ describe("Remote debug Tests", function () {
       );
       await driver.sleep(Timeout.longTimeWait);
       if (isRealKey) {
-        await validateWelcomeAndReplyBot(page, {
+        await validateCustomapi(page, {
           hasWelcomeMessage: false,
           hasCommandReplyValidation: true,
-          botCommand: "Tell me about Contoso Electronics history",
+          botCommand: "Get repairs for Karin",
           expectedWelcomeMessage: ValidationContent.AiChatBotWelcomeInstruction,
-          expectedReplyMessage: "1985",
+          expectedReplyMessage: "assignedTo: Karin",
           timeout: Timeout.longTimeWait,
         });
       } else {
-        await validateWelcomeAndReplyBot(page, {
+        await validateCustomapi(page, {
           hasWelcomeMessage: false,
           hasCommandReplyValidation: true,
           botCommand: "helloWorld",
