+++ packages/tests/src/unused-ui-test/remotedebug/remotedebug-tab-nosso-ts-win-only.test.ts
@@ -12,7 +12,7 @@ import {
   setSkuNameToB1,
   provisionProject,
   deployProject,
-} from "./remotedebugContext";
+} from "../../ui-test/remotedebug/remotedebugContext";
 import {
   execCommandIfExist,
   createNewProject,
